Inventory: movimientoTipo, ajuste UI, DB migration
Add support for movimientoTipo in inventory adjustments across frontend and backend. Frontend: extend InAjusteInventario API (movimientoTipoId, search with paging), add InMovimientoTipo client, implement AjusteBuscarModal and update AjusteInventarioView to select motivo, enforce CR rules, load lots/stock and show historial with pagination. Backend: update controllers, DTOs, services, JPA entities and repositories to handle movimiento types; add alert entities, event/listener, scheduler and a DB trigger. Also add a DB migration to convert various cantidad columns from DOUBLE PRECISION to INTEGER.

+++ eFacturador/src/main/java/com/braintech/eFacturador/dto/inventario/InAjusteInventarioDetalleRequestDTO.java
@@ -8,8 +8,8 @@ public class InAjusteInventarioDetalleRequestDTO {
   private String lote;
 
   /** Stock actual que el frontend leyó antes de enviar (para auditoría). */
-  private Double cantidadActual;
+  private Integer cantidadActual;
 
   /** Nuevo stock deseado. */
-  private Double cantidadNueva;
+  private Integer cantidadNueva;
 }
